# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'TelU Cup Backend API'
  description: 'Dokumentasi lengkap API manajemen kompetisi internal Tel-U Cup. Mencakup autentikasi, manajemen cabang olahraga, kontingen, registrasi tim, dan bagan pertandingan gugur.'
  version: 2.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'Development Server'
paths:
  /api/admin/templates:
    get:
      tags:
        - Admin
      summary: 'Daftar template turnamen'
      description: 'Mengambil daftar template tahun turnamen yang tersedia.'
      operationId: getTemplates
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    id:
                      type: integer
                      example: 1
                    name:
                      type: string
                      example: 'Telucup 2026'
                    year:
                      type: integer
                      example: 2026
                    created_at:
                      type: string
                      format: date-time
                  type: object
      security:
        -
          bearerAuth: []
  /api/admin/schedules:
    get:
      tags:
        - Admin
      summary: 'Jadwal pertandingan'
      description: 'Mengambil daftar jadwal pertandingan berdasarkan tanggal dan cabang olahraga. Gunakan GET /api/bracket untuk tampilan bagan lengkap.'
      operationId: getAdminSchedules
      parameters:
        -
          name: date
          in: query
          required: false
          schema:
            type: string
            format: date
        -
          name: sport_id
          in: query
          required: false
          schema:
            type: integer
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchObject'
      security:
        -
          bearerAuth: []
  '/api/admin/users/{id}/promote-to-pic':
    put:
      tags:
        - Admin
      summary: 'Promosikan player menjadi PIC Kontingen'
      description: 'Mengubah role pengguna menjadi `pic_kontingen`. Untuk juga menugaskan ke kontingen tertentu, gunakan PUT /api/contingents/{id}/assign-pic.'
      operationId: promoteToPic
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Berhasil dipromosikan'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'User berhasil dipromosikan menjadi PIC Kontingen'
                  user:
                    $ref: '#/components/schemas/UserObject'
                type: object
        400:
          description: 'Tidak dapat mengubah role panitia'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
  /api/register:
    post:
      tags:
        - Authentication
      summary: 'Daftarkan akun panitia baru'
      description: 'Membuat akun baru dengan role `panitia`. Registrasi publik hanya untuk panitia — pemain didaftarkan oleh PIC/panitia via `POST /api/players`.'
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  example: 'Rifqi Setiawan'
                  maxLength: 255
                email:
                  type: string
                  format: email
                  example: rifqi@telkomuniversity.ac.id
                password:
                  type: string
                  format: password
                  example: rahasia123
                  minLength: 8
              type: object
      responses:
        201:
          description: 'Registrasi berhasil'
          content:
            application/json:
              schema:
                properties:
                  user:
                    $ref: '#/components/schemas/UserObject'
                  token:
                    description: 'Bearer token untuk digunakan di header Authorization'
                    type: string
                    example: 1|abcdefg1234567890
                type: object
        422:
          description: 'Validasi gagal'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'The email has already been taken.'
                  errors:
                    properties:
                      email:
                        type: array
                        items:
                          type: string
                          example: 'The email has already been taken.'
                    type: object
                type: object
  /api/login:
    post:
      tags:
        - Authentication
      summary: 'Login dan dapatkan Bearer token'
      description: 'Autentikasi pengguna dengan email dan password. Gunakan token yang dikembalikan sebagai `Authorization: Bearer <token>` pada request selanjutnya.'
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: rifqi@telkomuniversity.ac.id
                password:
                  type: string
                  format: password
                  example: rahasia123
              type: object
      responses:
        200:
          description: 'Login berhasil'
          content:
            application/json:
              schema:
                properties:
                  user:
                    $ref: '#/components/schemas/UserObject'
                  token:
                    description: 'Bearer token untuk digunakan di header Authorization'
                    type: string
                    example: 2|xyz9876543210
                type: object
        422:
          description: 'Kredensial tidak valid'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'The given data was invalid.'
                  errors:
                    properties:
                      email:
                        type: array
                        items:
                          type: string
                          example: 'Kredensial yang diberikan tidak cocok dengan data kami.'
                    type: object
                type: object
...
  /campaign/confirm:
    post:
      tags:
        - Campaign
      summary: 'Konfirmasi baca kampanye & Sesi'
      description: 'Menyimpan konfirmasi bahwa pengguna telah membaca kampanye.'
      operationId: confirmCampaign
      requestBody:
        required: false
        content:
          application/json:
            schema:
              properties:
                dont_show_again:
                  type: boolean
                  example: true
              type: object
      responses:
        200:
          description: 'Konfirmasi diterima'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Konfirmasi diterima, lanjut ke dashboard.'
                type: object
...
  /api/players:
    post:
      tags:
        - Players
      summary: 'Buat akun player baru'
      description: 'Admin, panitia, atau PIC kontingen membuat akun player baru. Dibuat dengan data minimal — profil lengkap diisi kemudian oleh player sendiri via `PATCH /api/player/profile`.'
      operationId: storePlayer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  example: 'Ahmad Fauzi'
                  maxLength: 255
                email:
                  type: string
                  format: email
                  example: ahmad@telkomuniversity.ac.id
                password:
                  type: string
                  format: password
                  example: rahasia123
                  minLength: 8
              type: object
      responses:
        200:
          description: 'Player berhasil dibuat'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Player and user account created successfully'
                  player:
                    $ref: '#/components/schemas/PlayerObject'
                  user:
                    $ref: '#/components/schemas/UserObject'
                type: object
        422:
          description: 'Validasi gagal — email sudah digunakan'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: 'Role tidak diizinkan'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: 'Server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
    ErrorResponse:
      description: 'Format respon error standar'
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: 'Pesan error yang menjelaskan penyebab kegagalan.'
      type: object
...
    UserObject:
      description: 'Data akun pengguna'
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Budi Santoso'
        email:
          type: string
          format: email
          example: budi@telkomuniversity.ac.id
        role:
          type: string
          example: player
          enum:
            - admin
            - panitia
            - pic_kontingen
            - player
        is_kacamata:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
...
    ContingentObject:
      description: 'Kontingen yang mewakili satu fakultas atau divisi'
      properties:
        id:
          type: integer
          example: 3
        name:
          type: string
          example: 'Fakultas Informatika'
        pic_user_id:
          type: integer
          example: 5
          nullable: true
        pic:
          oneOf:
            -
              $ref: '#/components/schemas/UserObject'
          nullable: true
        players_count:
          description: 'Hanya tersedia pada endpoint list'
          type: integer
          example: 12
        created_at:
          type: string
          format: date-time
      type: object
    PlayerObject:
      description: 'Data pemain (athlete)'
      properties:
        id:
          type: integer
          example: 7
        user_id:
          type: integer
          example: 10
        name:
          type: string
          example: 'Ahmad Fauzi'
        nim_nip:
          type: string
          example: '1301234567'
          nullable: true
        sport_id:
          type: integer
          example: 1
          nullable: true
        sport_category_id:
          type: integer
          example: 2
          nullable: true
        contingent_id:
          type: integer
          example: 3
          nullable: true
        photo_path:
          type: string
          example: 'https://res.cloudinary.com/...'
          nullable: true
        verification_status:
          type: string
          example: verified
          nullable: true
        checked_in_at:
          type: string
          format: date-time
          nullable: true
        employee_status:
          type: string
          example: Mahasiswa
          nullable: true
        work_location:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      type: object
...
    TeamInMatch:
      description: 'Tim yang bertanding dalam satu pertandingan'
      properties:
        registration_id:
          type: integer
          example: 4
        contingent:
          $ref: '#/components/schemas/ContingentObject'
        players:
          type: array
          items:
            $ref: '#/components/schemas/PlayerObject'
      type: object
    MatchObject:
      description: 'Pertandingan dalam bagan gugur'
      properties:
        id:
          type: integer
          example: 12
        round:
          type: integer
          example: 2
        round_name:
          type: string
          example: Semifinal
        match_number:
          type: integer
          example: 1
        status:
          type: string
          enum:
            - scheduled
            - live
            - finished
            - bye
        match_date:
          type: string
          format: date
          example: '2026-06-15'
          nullable: true
        match_time:
          type: string
          example: '09:00'
          nullable: true
        location:
          type: string
          example: 'Gedung Sport Center Lt. 2'
          nullable: true
        referee_name:
          type: string
          example: 'Budi Santoso'
          nullable: true
        notes:
          type: string
          nullable: true
        score_a:
          type: integer
          example: 3
        score_b:
          type: integer
          example: 1
        team_a:
          oneOf:
            -
              $ref: '#/components/schemas/TeamInMatch'
          nullable: true
        team_b:
          oneOf:
            -
              $ref: '#/components/schemas/TeamInMatch'
          nullable: true
        winner:
          properties:
            registration_id:
              type: integer
            contingent:
              $ref: '#/components/schemas/ContingentObject'
          type: object
          nullable: true
        next_match_id:
          type: integer
          example: 15
          nullable: true
        next_match_slot:
          type: string
          nullable: true
          enum:
            - a
            - b
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...